ClipCascade_Desktop/clipcascade_linux_terminal/src/requirements.txt
pycryptodome==3.20.0
pyfiglet==1.0.2
pyperclip==1.9.0
Requests==2.32.3
websocket_client==1.8.0
xxhash==3.5.0
